fix: 렌더링 오류 수정
