+++ hanarowa/src/components/lesson/index.ts
@@ -1,6 +1,7 @@
 import CategoryFilter from './CategoryFilter';
 import Dropdown from './Dropdown';
 import LessonCard from './LessonCard';
+import LessonContent from './LessonContent';
 import LessonReservationCard from './LessonReservationCard';
 import StarRating from './StarRating';
 import StatusTab from './StatusTab';
@@ -9,6 +10,7 @@ export * from './detail';
 
 export {
   LessonCard,
+  LessonContent,
   LessonReservationCard,
   Dropdown,
   StarRating,
